reorganize sampling code.
